implementation du shema vu en cours ce qui donne : certaine competences peuvent avoir besoin d'un mode d'attaque(qu'elle vont devoir suivre) , on a une forge a projectile ( la boule de feu de salmeche fonctionne) , les projectile possede desormais les forge a effet pour creer les effets seulement lorsqu'il touche leur cible todo : effet ralentissement selon le temps + effet immobilisation

+++ src/main/java/fr/montreuil/iut/kalos_pokemon/Donne/Type.java
@@ -6,7 +6,7 @@ public enum Type {
     private static double superEfficace = 1.3;
 
 
-    public double calculDegats(Type typeDefenseur, int degats){
+    public double calculDegats(Type typeDefenseur, double degats){
 
         if (this == typeDefenseur || this == Type.neutre ||
             typeDefenseur == Type.neutre)
